feat(typography): allow ref pass through on Text components
Allow consumers of the `Text` componnent to pass down a `ref` to access the underlying DOM
element directly.

GitHub issue:

Clubhouse story:
https://app.clubhouse.io/czi-edu/story/141247/add-forward-ref-to-heading-text-components

I added a unit test that uses a ref to focus on the text element produced by a <Text/>
element to verify the ref is correctly being passed down.