feat(colors): add level 800 for alert and brand hues
We want a darker shade for the hues being used on the button color variants
which will then be used for the "active" button states. We're going to add
those colors as level 800.

So far we only have 2 confirmed colors here (brand 800 and alert 800), so
we're just going to add those for now and come back for the others later.