chore: use "workspace ranges" to specify dependencies between packages
To work around https://github.com/lerna/lerna/issues/1171.

The lerna publishing commands don't seem to update the yarn.lock file,
unfortunately. This means on CI when we run `yarn install --immutable`
we get an error because yarn tries to make changes to the lockfile
(which lerna should have, but didn't, make).

If we specify with dependencies between workspace packages with
https://yarnpkg.com/features/workspaces/#workspace-ranges-workspace, then
everything works because the version number of the packages doesn't actually
appear in the yarn.lock file, and therefore doesn't need to get updated.