test(storybook): use @storybook/testing-react for writing tests based on stories
Instead of our own `prepareStory`, now that Storybook has released an official utility.

Note that @storybook/testing-react does not currently support passing refs to components using
`React.forwardRef`. This makes sense, because what's being rendered is the story (which does not
accept refs), not the component (which does). In any case, we work around it here by directly
using the actual component in the 2 tests relying on refs.