feat: export ClickableStyleProps from the top-level index.ts
Usages in our Remix apps will not need to deeply import anymore (in fact,
they cannot with Vite due to how the bundling works)

  - import type {ClickableStyleProps} from "@chanzuckerberg/eds/lib/components/ClickableStyle";
  + import type {ClickableStyleProps} from "@chanzuckerberg/eds";

We need the ClickableStyleProps for our custom Remix Link component, so
we properly integrate the Remix and EDS Link components.

Re-exporting the props is easier to work with. The deep import in edu-stack was always a hack.