feat(Tooltip)!: remove deprecated props (#1861)
- remove `align` prop, in place of the better-named `placement` prop

This has been marked as deprecated for a while now, so cleaning up the
code associated with this. In order to address in a consuming project,
you can replace `align` with `placement` in all use cases (this was
happening internally already).

- remove `variant` prop

The prop was used for a partially-implemented handling of light/dark
mode, and had been marked as deprecated. Removing now, making usage of
the light token values as the current, and only, option.

- clean up documentation, stoybook config, and test snapshots